benchmark_suit/requirements.txt
numa==1.4.4
PyYAML==3.12
Yapsy==1.11.223
py-cpuinfo==8.0.0
